requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
whitenoise==5.3.0
